Move MarkdownFileManager to parsing package
